feat: Add support for the Communications MFE
The MFE is usable by instructors to send bulk email to learners in a course.

The feature itself (the ability to send bulk email) pre-dates this MFE, and
must be enabled as usual for the "Email" tab to become visible in the
Instructor Dashboard.  The difference is that with this change, the tab will
include a link to the MFE by default.